+++ src/components/Contacts/ContactDetails/ContactDetailsTab/ContactDetailsTab.stories.tsx
@@ -1,7 +1,7 @@
 import React, { ReactElement } from 'react';
 import { GqlMockedProvider } from '../../../../../__tests__/util/graphqlMocking';
-import { ContactDetailTab } from './ContactDetailTab';
-import { ContactDetailTabQuery } from './ContactDetailTab.generated';
+import { ContactDetailTab } from './ContactDetailsTab';
+import { ContactDetailsTabQuery } from './ContactDetailsTab.generated';
 
 export default {
   title: 'Contacts/Tab/ContactDetailTab',
@@ -13,7 +13,7 @@ const contactId = '222';
 
 export const Default = (): ReactElement => {
   return (
-    <GqlMockedProvider<ContactDetailTabQuery>>
+    <GqlMockedProvider<ContactDetailsTabQuery>>
       <ContactDetailTab accountListId={accountListId} contactId={contactId} />
     </GqlMockedProvider>
   );
